hardening(codeql): resolve all open code scanning alerts
CodeQL flagged one critical and twenty-one high findings plus a missing workflow permission. All are fixed with no behavior change: created files still get mode 0644 (the VFS default the old zero-mode path sanitized to) and shell/WM output is unchanged.

The pipe readiness check compared an unsigned difference to zero; it now compares the byte count against PIPE_BUFFER_SIZE directly. Size multiplications fed to malloc/realloc/memmove/memset in fat32, kernel terminal, clock, calendar, dock, and the terminal app are widened to size_t explicitly so the 32-bit product cannot wrap. The xHCI protocol-capability walk uses a 32-bit counter bounded by the port_protocols table instead of a uint8_t that could wrap or run past the array.

The libc open() shim dropped the creation mode even though sys_open already honors it, so it is now variadic and forwards the POSIX mode argument; every O_CREAT call site passes 0644, and latitude's raw-syscall workaround is retired. TOCTOU findings are resolved by opening first and checking after: wav.c and the shell's get_file_data size files with fsize(fd), ls/cp/du stat only after a successful open, and the WM storage bootstrap probes the target with a read-only open instead of stat before create. ci.yml gains top-level permissions: contents: read.

+++ src/fs/fat32/fat32.cpp
@@ -343,7 +343,7 @@ static bool dir_chain_load(FAT32Filesystem *fs, uint32_t dir_cluster, DirChain *
         return false;
 
     out->clusters = static_cast<uint32_t *>(malloc(sizeof(uint32_t) * cluster_count));
-    out->data = static_cast<uint8_t *>(malloc(out->cluster_size * cluster_count));
+    out->data = static_cast<uint8_t *>(malloc((size_t)out->cluster_size * cluster_count));
     if (!out->clusters || !out->data) {
         free(out->clusters);
         free(out->data);
@@ -429,7 +429,7 @@ static bool dir_chain_extend(DirChain *chain, uint32_t required_entries)
         chain->clusters = new_clusters;
 
         uint8_t *new_data =
-            static_cast<uint8_t *>(realloc(chain->data, chain->cluster_size * (chain->cluster_count + 1)));
+            static_cast<uint8_t *>(realloc(chain->data, (size_t)chain->cluster_size * (chain->cluster_count + 1)));
         if (!new_data)
             return false;
         chain->data = new_data;
